Initial commit: working FsTree impl
This is my "hello shelly" app, learning how to use shelly and
related haskell libraries to do scripting-like tasks.

At the moment what I have are a bunch of helper functions to work
with FilePaths without doing much IO, and a directory tree
implementation; the general goal is to get something similar to
ghar (a now-dead python tool that would symlink files in the home
directory to files in repositories) working in haskell.